重複している行を共通にした
