+++ app/src/main/res/values/strings.xml
@@ -89,4 +89,5 @@
     <string name="amazonAdd">アマゾンから追加</string>
     <string name="delete_item">項目を削除</string>
     <string name="NC_transition">NC_transition</string>
+    <string name="Error">エラーが発生しました</string>
 </resources>
